feat: finish v1
