+++ .github/workflows/release.yml
@@ -6,6 +6,16 @@ on:
       - master
 
 jobs:
+  setup-environment:
+    name: Setup environment
+    runs-on: ubuntu-latest
+    timeout-minutes: 1
+    steps:
+      - name: Installing node.js
+        uses: actions/setup-node@v1
+        with:
+          node-version: 14.6.0
+
   lint:
     name: ESLint
     runs-on: ubuntu-latest
@@ -23,12 +33,35 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-yarn-
       - name: Installing dependencies
+        if: steps.cache-modules.outputs.cache-hit != 'true'
         run: yarn --frozen-lockfile
       - run: yarn lint:code
+
+  test:
+    name: Test
+    runs-on: ubuntu-latest
+    timeout-minutes: 2
+    steps:
+      - name: Get Yarn cache path
+        id: yarn-cache
+        run: echo "::set-output name=dir::$(yarn cache dir)"
+      - uses: actions/checkout@master
+      - name: Load Yarn cache
+        uses: actions/cache@master
+        with:
+          path: ${{ steps.yarn-cache.outputs.dir }}
+          key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-yarn-
+      - name: Installing dependencies
+        if: steps.cache-modules.outputs.cache-hit != 'true'
+        run: yarn --frozen-lockfile
+      - run: yarn test:coverage --ci
 
   build-publish:
     name: Build and Publish
     runs-on: ubuntu-latest
+    needs: [lint, test]
     steps:
       - uses: actions/checkout@master
       - name: Load Yarn cache
@@ -40,9 +73,8 @@ jobs:
       - name: Build
         if: steps.cache-modules.outputs.cache-hit != 'true'
         run: yarn --frozen-lockfile
-      - name: Publish
-        if: github.ref == 'refs/heads/master'
-        uses: mikeal/merge-release@master
+      - name: Release
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          NPM_AUTH_TOKEN: ${{ secrets.npm_token }}
+          NPM_TOKEN: ${{ secrets.npm_token }}
+        run: yarn semantic-release
